:recycle: refactor: fetch a users's current course from the learners service
