+++ libs/features/convs-mgr/learners/src/lib/pages/learners-page/learners-page.component.ts
@@ -6,12 +6,10 @@ import { LiveAnnouncer } from '@angular/cdk/a11y';
 import { SelectionModel } from '@angular/cdk/collections';
 
 import { SubSink } from 'subsink';
-import { map, switchMap, of, combineLatest } from 'rxjs';
 
 import { EnrolledEndUser, EnrolledEndUserStatus } from '@app/model/convs-mgr/learners';
 
 import { EnrolledLearnersService } from '@app/state/convs-mgr/learners';
-import { EndUserService } from '@app/state/convs-mgr/end-users';
 
 import { BulkActionsModalComponent } from '../../modals/bulk-actions-modal/bulk-actions-modal.component';
 
@@ -40,42 +38,19 @@ export class LearnersPageComponent implements OnInit, OnDestroy {
 
   constructor(
     private _eLearners: EnrolledLearnersService,
-    private _endUsers: EndUserService,
     private _liveAnnouncer: LiveAnnouncer,
     private _dialog: MatDialog
   ) {}
 
   ngOnInit() {
-    this.getCurrentCourseForUsers();
+    this.getLearners();
     this.getAllClasses();
     this.getAllCourses();
     this.getAllPlatforms();
   }
 
-  // TODO @LemmyMwaura: We now update the enrolled user's current milestone from the event brick so this method is redundant.
-  // TODO: to be removed after incremental adoption of the event brick(as a way to mark entry into a new milstone) on existing prod versions.
-  /** Get's the current course for all user's with the WhatsappEndUser Id  */
-  getCurrentCourseForUsers() {
-    const allLearners$ = this._eLearners.getAllLearners$().pipe(
-      switchMap((enrolledUsrs) => {
-        if (enrolledUsrs.length === 0) return of([]);
-
-        const endUsers = enrolledUsrs.map((user) => {
-          if (!user.whatsappUserId) return of(user);
-
-          return this._endUsers.getCourse(user.whatsappUserId).pipe(
-            map((eventStack) => {
-              if (eventStack && eventStack[0].isMilestone) {
-                user.currentCourse = eventStack[0].name;
-              }
-              return user;
-            })
-          );
-        });
-
-        return combineLatest(endUsers);
-      })
-    );
+  getLearners() {
+    const allLearners$ = this._eLearners.getAllLearners$();
 
     this._sBs.sink = allLearners$.subscribe((alllearners) => {
       this.dataSource.data = alllearners;
